Ask for a bigger buffer instead of panicking
This loosely starts to set errnop and h_errnop but has not been
done thoroughly or necessarily accurately.

The repeated pattern of matching the translation into a C struct in the
buffer seems pretty indicative that it should be factored out or pushed
into a common trait. I have some ideas for this but wanted to start with
the repetitive version as a proof-of-concept.