Allow hooks to set NSS Status
This switches away from the hooks using a simple Option type to one that
more closely reflects the underlying APIs, allowing the conveyance of
different types of failures.

Since this introduces an enum which holds any relevant returned value,
we can also consolidate a lot of the "to_c" logic into a trait and avoid
some repitition.

This should also be a pathway to more safety.